backend/requirements.txt
# FastAPI & server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# ML & Audio
torch>=2.2.0
numpy>=1.26.0
Pillow>=10.0.0
librosa>=0.10.0
soundfile>=0.12.0

# StyleGAN3 dependencies
scipy>=1.12.0
requests>=2.31.0
click>=8.1.0
ninja>=1.11.0
